+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "tidebit-defi",
-  "version": "0.9.0+21",
+  "version": "0.9.0+22",
   "private": true,
   "scripts": {
     "dev": "next dev",
