+++ tailwind.config.ts
@@ -683,7 +683,7 @@ module.exports = {
         'tabs-text-active-primary': 'var(--color-tabs-text-active-primary)',
         'tabs-text-default': 'var(--color-tabs-text-default)',
         'tabs-text-disable': 'var(--color-tabs-text-disable)',
-        'tabs-text-hove-neutral': 'var(--color-tabs-text-hove-neutral)',
+        'tabs-text-hover-neutral': 'var(--color-tabs-text-hover-neutral)',
         'tabs-text-hover-primary': 'var(--color-tabs-text-hover-primary)',
         'text-brand-on-primary-container': 'var(--color-text-brand-on-primary-container)',
         'text-brand-on-primary': 'var(--color-text-brand-on-primary)',
